requirements.txt
absl-py==1.4.0
altair==5.4.1
asttokens
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama
comm
contourpy==1.3.0
cycler==0.12.1
debugpy
decorator
dm-tree==0.1.8
exceptiongroup
executing
filelock==3.13.1
fonttools==4.53.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.8
importlib_metadata
ipykernel
ipython
jedi
Jinja2==3.1.3
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client
jupyter_core
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.6.1
nest_asyncio
networkx==3.2.1
numpy==1.26.3
packaging
pandas==2.2.2
parso
pickleshare
pillow==10.2.0
platformdirs
prompt_toolkit
protobuf==5.28.0
psutil
pure_eval
pyarrow==17.0.0
pydeck==0.9.1
Pygments
pyparsing==3.1.4
python-dateutil
pytz==2024.1
pywin32==306
pyzmq
referencing==0.35.1
requests==2.32.3
rich==13.8.0
rpds-py==0.20.0
six
smmap==5.0.1
stack-data
streamlit==1.38.0
sympy==1.12
tenacity==8.5.0
tensorflow-model-optimization==0.8.0
toml==0.10.2
torch==2.4.0+cu124
torchaudio==2.4.0+cu124
torchvision==0.19.0+cu124
tornado
tqdm==4.66.5
traitlets
typing_extensions
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.2
wcwidth
zipp
